added ConvexGeometry, ConvexGeometry and BufferGeometryUtils

+++ core/src/main/java/org/treblereel/gwt/three4g/core/BufferAttribute.java
@@ -179,9 +179,9 @@ public class BufferAttribute {
   }
 
   public JsArray<Double> array;
-  public double count;
+  public int count;
   public boolean isBufferAttribute;
-  public double itemSize;
+  public int itemSize;
   public String name;
   public boolean normalized;
   public BufferAttribute.OnUploadCallbackFn0 onUploadCallback;
